+++ src/types/hackcraft2.d.ts
@@ -1,9 +1,11 @@
-export declare interface Hackcraf2Settings {
+export declare interface Hackcraft2Settings {
     serverAddress: string | undefined;
     playerId: string | undefined;
+    serverAddressHistory: string | undefined;
+    playerIdHistory: string | undefined;
 }
 
-export declare interface Hackcraf2SourceFile {
+export declare interface Hackcraft2SourceFile {
     fileName: string | undefined;
     languageId: string | undefined;
     code: string | undefined;