+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
-# 0.0.1 2023-4-9
+# 0.0.2 2023-4-10
 
-- Init.
+- Future: Server Address History.
+- Future: Player Id History.
+
+# 0.0.1 2023-4-10
+
+- Preview Release.